Project 1 for class
The first programming project involves writing a program that evaluates infix expressions of
unsigned integers using two stacks. The program should consist of three classes. The main class
should create a GUI that allows the user input an infix expression and displays the result. The
GUI should look as follows: